# ui/valset-standardizer-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:30066"
    }
  ],
...
  "paths": {
...
    "/api/valuation-workflows/upload": {
      "post": {
        "summary": "上传外部估值表并完成 ODS 原始提取。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultUploadValuationFileResponse"
                }
              }
            }
          }
        },
        "operationId": "upload",
        "parameters": []
      }
    },
    "/api/valuation-workflows/analyze": {
      "post": {
        "summary": "执行 DWD 外部估值解析。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataSourceType": {
                    "format": "string",
                    "description": "估值表数据源类型 (EXCEL, CSV, API, DB)，默认 EXCEL 保持向后兼容。",
                    "type": "string"
                  },
                  "workbookPath": {
                    "format": "string",
                    "description": "要解析的数据源URI或绝对路径。",
                    "type": "string"
                  },
                  "forceRebuild": {
                    "format": "boolean",
                    "description": "是否强制重新生成解析任务。",
                    "type": "boolean",
                    "example": true
                  },
                  "fileId": {
                    "format": "int64",
                    "description": "可选的原始数据文件标识。",
                    "type": "number",
                    "example": 0
                  },
                  "fileNameOriginal": {
                    "format": "string",
                    "description": "可选的原始文件名。",
                    "type": "string"
                  },
                  "createdBy": {
                    "format": "string",
                    "description": "可选的创建者标识符。",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTaskViewDTO"
                }
              }
            }
          }
        },
        "operationId": "analyze",
        "parameters": []
      }
    },
    "/api/valuation-workflows/match": {
      "post": {
        "summary": "执行外部估值科目匹配。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataSourceType": {
                    "format": "string",
                    "description": "估值表数据源类型 (EXCEL, CSV, API, DB)，默认 EXCEL 保持向后兼容。",
                    "type": "string"
                  },
                  "workbookPath": {
                    "format": "string",
                    "description": "评估估值表的数据源URI或绝对路径。",
                    "type": "string"
                  },
                  "forceRebuild": {
                    "format": "boolean",
                    "description": "是否强制重新生成匹配任务。",
                    "type": "boolean",
                    "example": true
                  },
                  "fileId": {
                    "format": "int64",
                    "description": "可选的原始数据文件标识。",
                    "type": "number",
                    "example": 0
                  },
                  "topK": {
                    "format": "int32",
                    "description": "每个科目保留的考生数量。",
                    "type": "integer",
                    "example": 0
                  },
                  "createdBy": {
                    "format": "string",
                    "description": "可选的创建者标识符。",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTaskViewDTO"
                }
              }
            }
          }
        },
        "operationId": "match",
        "parameters": []
      }
    },
    "/api/valuation-workflows/full-process": {
      "post": {
        "summary": "上传文件并串联执行提取、解析、匹配。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultFullWorkflowResponse"
                }
              }
            }
          }
        },
        "operationId": "fullProcess",
        "parameters": []
      }
    },
    "/api/valuation-workflows/{fileId}/raw-data": {
      "get": {
        "summary": "查询 ODS 原始行数据。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultRawValuationDataViewDTO"
                }
              }
            }
          }
        },
        "operationId": "queryRawData",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "文件主键",
            "required": true,
            "example": 0,
            "name": "fileId"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
    "/api/valuation-workflows/{fileId}/stg-data": {
      "get": {
        "summary": "查询 STG 外部估值解析快照。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultStgExternalValuationViewDTO"
                }
              }
            }
          }
        },
        "operationId": "queryStgData",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "文件主键",
            "required": true,
            "example": 0,
            "name": "fileId"
          }
        ]
      }
    },
    "/api/valuation-workflows/{fileId}/dwd-data": {
      "get": {
        "summary": "查询 DWD 外部估值标准数据。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultDwdExternalValuationViewDTO"
                }
              }
            }
          }
        },
        "operationId": "queryDwdData",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "文件主键",
            "required": true,
            "example": 0,
            "name": "fileId"
          }
        ]
      }
    },
    "/api/valuation-workflows/{fileId}/match-results": {
      "get": {
        "summary": "查询匹配结果。",
        "deprecated": false,
        "tags": [
          "ValuationWorkflowController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultMatchResultViewDTO"
                }
              }
            }
          }
        },
        "operationId": "queryMatchResults",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "文件主键",
            "required": true,
            "example": 0,
            "name": "fileId"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SingleResultStgExternalValuationViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/StgExternalValuationViewDTO"
          }
        }
      },
...
      "defaultPrimitive": {
        "type": "object",
        "properties": {
          " -": {
            "description": "array of string"
          }
        }
      },
      "SingleResultTaskViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TaskViewDTO"
          }
        }
      },
...
      "Object": {
        "type": "object",
        "properties": {
          "mapKey": {
            "description": "A map key.(object)",
            "$ref": "#/components/schemas/Object"
          }
        }
      },
...
      "SingleResultFullWorkflowResponse": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/FullWorkflowResponse"
          }
        }
      },
...
      "SingleResultDwdExternalValuationViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/DwdExternalValuationViewDTO"
          }
        }
      },
...
      "DwdExternalValuationViewDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "workbookPath": {
            "format": "string",
            "description": "工作簿路径。",
            "type": "string"
          },
          "sheetName": {
            "format": "string",
            "description": "工作表名称。",
            "type": "string"
          },
          "headerRowNumber": {
            "format": "int32",
            "description": "表头行号。",
            "type": "integer",
            "example": 0
          },
          "dataStartRowNumber": {
            "format": "int32",
            "description": "数据起始行号。",
            "type": "integer",
            "example": 0
          },
          "fileNameOriginal": {
            "format": "string",
            "description": "原始文件名。",
            "type": "string"
          },
          "title": {
            "format": "string",
            "description": "标题。",
            "type": "string"
          },
          "basicInfo": {
            "description": "基础信息。(object)",
            "$ref": "#/components/schemas/String"
          },
          "headers": {
            "description": "表头内容。",
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": "\"\"\",\"\"\""
          },
          "headerDetails": {
            "description": "表头明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/defaultPrimitive"
            }
          },
          "headerColumns": {
            "description": "表头元数据。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeaderColumnMeta"
            }
          },
          "subjects": {
            "description": "科目明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectRecord"
            }
          },
          "metrics": {
            "description": "指标明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricRecord"
            }
          }
        }
      },
      "SingleResultRawValuationDataViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/RawValuationDataViewDTO"
          }
        }
      },
...
      "TaskViewDTO": {
        "type": "object",
        "properties": {
          "taskId": {
            "format": "string",
            "description": "唯一的任务标识符。",
            "type": "string"
          },
          "taskType": {
            "format": "string",
            "description": "任务类型名称。",
            "type": "string"
          },
          "taskStage": {
            "format": "string",
            "description": "任务阶段标志。",
            "type": "string"
          },
          "taskStatus": {
            "format": "string",
            "description": "当前任务状态。",
            "type": "string"
          },
          "businessKey": {
            "format": "string",
            "description": "创建任务时使用的业务密钥。",
            "type": "string"
          },
          "inputPayload": {
            "format": "string",
            "description": "原始输入负载存储在任务记录中。",
            "type": "string"
          },
          "inputData": {
            "description": "当输入负载为有效的JSON时，已解析输入负载。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "resultPayload": {
            "format": "string",
            "description": "原始结果负载存储在任务记录中。",
            "type": "string"
          },
          "resultData": {
            "description": "当结果有效负载为有效JSON时，解析结果负载。(object)",
            "$ref": "#/components/schemas/Object"
          },
          "errorMessage": {
            "format": "string",
            "description": "任务失败时的可读原因，优先取结果负载里的 errorMessage。",
            "type": "string"
          },
          "errorCode": {
            "format": "string",
            "description": "任务失败时的错误分类码，供前端按固定状态展示。",
            "type": "string"
          },
          "rowCount": {
            "format": "string",
            "description": "提取任务的总行数。",
            "type": "string"
          },
          "fileSizeBytes": {
            "format": "string",
            "description": "提取任务的源文件大小，单位字节。",
            "type": "string"
          },
          "durationMs": {
            "format": "string",
            "description": "提取任务耗时，单位毫秒。",
            "type": "string"
          },
          "taskStartTime": {
            "format": "string",
            "description": "任务开始时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          },
          "parseTaskTimeMs": {
            "format": "string",
            "description": "文件解析耗时，单位毫秒。",
            "type": "string"
          },
          "standardizeTimeMs": {
            "format": "string",
            "description": "结构标准化耗时，单位毫秒。",
            "type": "string"
          },
          "matchStandardSubjectTimeMs": {
            "format": "string",
            "description": "匹配标准科目耗时，单位毫秒。",
            "type": "string"
          }
        }
      },
...
      "SingleResultMatchResultViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/MatchResultViewDTO"
          }
        }
      },
...
      "RawValuationDataViewDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "totalRows": {
            "format": "int32",
            "description": "原始行总数。",
            "type": "integer",
            "example": 0
          },
          "sheets": {
            "description": "原始 sheet 列表。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RawValuationSheetDTO"
            }
          },
          "rows": {
            "description": "原始行明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RawValuationRowDTO"
            }
          }
        }
      },
...
      "MatchResultViewDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "matchedCount": {
            "format": "int32",
            "description": "已匹配条数。",
            "type": "integer",
            "example": 0
          },
          "results": {
            "description": "匹配结果列表。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValsetMatchResult"
            }
          }
        }
      },
...
      "StgExternalValuationViewDTO": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "workbookPath": {
            "format": "string",
            "description": "工作簿路径。",
            "type": "string"
          },
          "sheetName": {
            "format": "string",
            "description": "工作表名称。",
            "type": "string"
          },
          "headerRowNumber": {
            "format": "int32",
            "description": "表头行号。",
            "type": "integer",
            "example": 0
          },
          "dataStartRowNumber": {
            "format": "int32",
            "description": "数据起始行号。",
            "type": "integer",
            "example": 0
          },
          "fileNameOriginal": {
            "format": "string",
            "description": "原始文件名。",
            "type": "string"
          },
          "title": {
            "format": "string",
            "description": "标题。",
            "type": "string"
          },
          "basicInfo": {
            "description": "基础信息。(object)",
            "$ref": "#/components/schemas/String"
          },
          "headers": {
            "description": "表头内容。",
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": "\"\"\",\"\"\""
          },
          "headerDetails": {
            "description": "表头明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/defaultPrimitive"
            }
          },
          "headerColumns": {
            "description": "表头元数据。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeaderColumnMeta"
            }
          },
          "subjects": {
            "description": "科目明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectRecord"
            }
          },
          "metrics": {
            "description": "指标明细。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MetricRecord"
            }
          }
        }
      },
...
      "FullWorkflowResponse": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "workbookPath": {
            "format": "string",
            "description": "工作簿路径。",
            "type": "string"
          },
          "dataSourceType": {
            "format": "string",
            "description": "数据源类型。",
            "type": "string"
          },
          "fileFingerprint": {
            "format": "string",
            "description": "文件指纹。",
            "type": "string"
          },
          "filesysTaskId": {
            "format": "string",
            "description": "文件服务任务标识。",
            "type": "string"
          },
          "filesysFileId": {
            "format": "string",
            "description": "文件服务文件标识。",
            "type": "string"
          },
          "filesysObjectKey": {
            "format": "string",
            "description": "文件服务对象键。",
            "type": "string"
          },
          "filesysInstantUpload": {
            "format": "boolean",
            "description": "是否立即完成文件服务上传。",
            "type": "boolean",
            "example": true
          },
          "extractTask": {
            "description": "提取任务详情。(object)",
            "$ref": "#/components/schemas/TaskViewDTO"
          },
          "parseTask": {
            "description": "解析任务详情。(object)",
            "$ref": "#/components/schemas/TaskViewDTO"
          },
          "matchTask": {
            "description": "匹配任务详情。(object)",
            "$ref": "#/components/schemas/TaskViewDTO"
          }
        }
      },
...
      "SingleResultUploadValuationFileResponse": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/UploadValuationFileResponse"
          }
        }
      },
...
      "UploadValuationFileResponse": {
        "type": "object",
        "properties": {
          "fileId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "workbookPath": {
            "format": "string",
            "description": "工作簿路径。",
            "type": "string"
          },
          "dataSourceType": {
            "format": "string",
            "description": "数据源类型。",
            "type": "string"
          },
          "fileSizeBytes": {
            "format": "string",
            "description": "文件大小，单位字节。",
            "type": "string"
          },
          "fileFingerprint": {
            "format": "string",
            "description": "文件指纹。",
            "type": "string"
          },
          "filesysTaskId": {
            "format": "string",
            "description": "文件服务任务标识。",
            "type": "string"
          },
          "filesysFileId": {
            "format": "string",
            "description": "文件服务文件标识。",
            "type": "string"
          },
          "filesysObjectKey": {
            "format": "string",
            "description": "文件服务对象键。",
            "type": "string"
          },
          "filesysInstantUpload": {
            "format": "boolean",
            "description": "是否立即完成文件服务上传。",
            "type": "boolean",
            "example": true
          },
          "reusedExistingExtractTask": {
            "format": "boolean",
            "description": "是否复用已有提取任务。",
            "type": "boolean",
            "example": true
          },
          "extractTask": {
            "description": "提取任务详情。(object)",
            "$ref": "#/components/schemas/TaskViewDTO"
          }
        }
      },
...
    }
  }
}